refactor: query-builder: Move query builders to own directory
Moves QueryBuilder and its subclasses to their own directory.
