+++ src/metadata/RelationCountMetadata.ts
@@ -1,7 +1,7 @@
 import {RelationCountMetadataArgs} from "../metadata-args/RelationCountMetadataArgs";
 import {EntityMetadata} from "./EntityMetadata";
 import {RelationMetadata} from "./RelationMetadata";
-import {SelectQueryBuilder} from "../query-builder/SelectQueryBuilder";
+import {SelectQueryBuilder} from "../query-builder/builder/SelectQueryBuilder";
 
 /**
  * Contains all information about entity's relation count.
