+++ src/query-builder/builder/DeleteQueryBuilder.ts
@@ -1,22 +1,22 @@
-import {CockroachDriver} from "../driver/cockroachdb/CockroachDriver";
-import {OracleDriver} from "../driver/oracle/OracleDriver";
+import {CockroachDriver} from "../../driver/cockroachdb/CockroachDriver";
+import {OracleDriver} from "../../driver/oracle/OracleDriver";
 import {QueryBuilder} from "./QueryBuilder";
-import {ObjectLiteral} from "../common/ObjectLiteral";
-import {EntityTarget} from "../common/EntityTarget";
-import {Connection} from "../connection/Connection";
-import {QueryRunner} from "../query-runner/QueryRunner";
-import {SqlServerDriver} from "../driver/sqlserver/SqlServerDriver";
-import {PostgresDriver} from "../driver/postgres/PostgresDriver";
-import {WhereExpression} from "./WhereExpression";
-import {Brackets} from "./Brackets";
-import {DeleteResult} from "./result/DeleteResult";
-import {ReturningStatementNotSupportedError} from "../error/ReturningStatementNotSupportedError";
-import {SqljsDriver} from "../driver/sqljs/SqljsDriver";
-import {MysqlDriver} from "../driver/mysql/MysqlDriver";
-import {BroadcasterResult} from "../subscriber/BroadcasterResult";
-import {EntitySchema} from "../index";
-import {AuroraDataApiDriver} from "../driver/aurora-data-api/AuroraDataApiDriver";
-import {BetterSqlite3Driver} from "../driver/better-sqlite3/BetterSqlite3Driver";
+import {ObjectLiteral} from "../../common/ObjectLiteral";
+import {EntityTarget} from "../../common/EntityTarget";
+import {Connection} from "../../connection/Connection";
+import {QueryRunner} from "../../query-runner/QueryRunner";
+import {SqlServerDriver} from "../../driver/sqlserver/SqlServerDriver";
+import {PostgresDriver} from "../../driver/postgres/PostgresDriver";
+import {WhereExpression} from "../WhereExpression";
+import {Brackets} from "../Brackets";
+import {DeleteResult} from "../result/DeleteResult";
+import {ReturningStatementNotSupportedError} from "../../error/ReturningStatementNotSupportedError";
+import {SqljsDriver} from "../../driver/sqljs/SqljsDriver";
+import {MysqlDriver} from "../../driver/mysql/MysqlDriver";
+import {BroadcasterResult} from "../../subscriber/BroadcasterResult";
+import {EntitySchema} from "../../index";
+import {AuroraDataApiDriver} from "../../driver/aurora-data-api/AuroraDataApiDriver";
+import {BetterSqlite3Driver} from "../../driver/better-sqlite3/BetterSqlite3Driver";
 
 /**
  * Allows to build complex sql queries in a fashion way and execute those queries.
