+++ src/query-builder/builder/InsertQueryBuilder.ts
@@ -1,24 +1,24 @@
-import {CockroachDriver} from "../driver/cockroachdb/CockroachDriver";
-import {SapDriver} from "../driver/sap/SapDriver";
+import {CockroachDriver} from "../../driver/cockroachdb/CockroachDriver";
+import {SapDriver} from "../../driver/sap/SapDriver";
 import {QueryBuilder} from "./QueryBuilder";
-import {ObjectLiteral} from "../common/ObjectLiteral";
-import {EntityTarget} from "../common/EntityTarget";
-import {QueryDeepPartialEntity} from "./QueryPartialEntity";
-import {SqlServerDriver} from "../driver/sqlserver/SqlServerDriver";
-import {PostgresDriver} from "../driver/postgres/PostgresDriver";
-import {MysqlDriver} from "../driver/mysql/MysqlDriver";
-import {RandomGenerator} from "../util/RandomGenerator";
-import {InsertResult} from "./result/InsertResult";
-import {ReturningStatementNotSupportedError} from "../error/ReturningStatementNotSupportedError";
-import {InsertValuesMissingError} from "../error/InsertValuesMissingError";
-import {ColumnMetadata} from "../metadata/ColumnMetadata";
-import {ReturningResultsEntityUpdator} from "./ReturningResultsEntityUpdator";
-import {AbstractSqliteDriver} from "../driver/sqlite-abstract/AbstractSqliteDriver";
-import {SqljsDriver} from "../driver/sqljs/SqljsDriver";
-import {BroadcasterResult} from "../subscriber/BroadcasterResult";
-import {EntitySchema} from "../entity-schema/EntitySchema";
-import {OracleDriver} from "../driver/oracle/OracleDriver";
-import {AuroraDataApiDriver} from "../driver/aurora-data-api/AuroraDataApiDriver";
+import {ObjectLiteral} from "../../common/ObjectLiteral";
+import {EntityTarget} from "../../common/EntityTarget";
+import {QueryDeepPartialEntity} from "../QueryPartialEntity";
+import {SqlServerDriver} from "../../driver/sqlserver/SqlServerDriver";
+import {PostgresDriver} from "../../driver/postgres/PostgresDriver";
+import {MysqlDriver} from "../../driver/mysql/MysqlDriver";
+import {RandomGenerator} from "../../util/RandomGenerator";
+import {InsertResult} from "../result/InsertResult";
+import {ReturningStatementNotSupportedError} from "../../error/ReturningStatementNotSupportedError";
+import {InsertValuesMissingError} from "../../error/InsertValuesMissingError";
+import {ColumnMetadata} from "../../metadata/ColumnMetadata";
+import {ReturningResultsEntityUpdator} from "../ReturningResultsEntityUpdator";
+import {AbstractSqliteDriver} from "../../driver/sqlite-abstract/AbstractSqliteDriver";
+import {SqljsDriver} from "../../driver/sqljs/SqljsDriver";
+import {BroadcasterResult} from "../../subscriber/BroadcasterResult";
+import {EntitySchema} from "../../entity-schema/EntitySchema";
+import {OracleDriver} from "../../driver/oracle/OracleDriver";
+import {AuroraDataApiDriver} from "../../driver/aurora-data-api/AuroraDataApiDriver";
 
 /**
  * Allows to build complex sql queries in a fashion way and execute those queries.
