+++ src/query-builder/builder/QueryBuilder.ts
@@ -1,28 +1,28 @@
-import {ObjectLiteral} from "../common/ObjectLiteral";
-import {QueryRunner} from "../query-runner/QueryRunner";
-import {Connection} from "../connection/Connection";
-import {QueryExpressionMap} from "./QueryExpressionMap";
+import {ObjectLiteral} from "../../common/ObjectLiteral";
+import {QueryRunner} from "../../query-runner/QueryRunner";
+import {Connection} from "../../connection/Connection";
+import {QueryExpressionMap} from "../QueryExpressionMap";
 import {SelectQueryBuilder} from "./SelectQueryBuilder";
 import {UpdateQueryBuilder} from "./UpdateQueryBuilder";
 import {DeleteQueryBuilder} from "./DeleteQueryBuilder";
 import {SoftDeleteQueryBuilder} from "./SoftDeleteQueryBuilder";
 import {InsertQueryBuilder} from "./InsertQueryBuilder";
 import {RelationQueryBuilder} from "./RelationQueryBuilder";
-import {EntityTarget} from "../common/EntityTarget";
-import {Alias} from "./Alias";
-import {Brackets} from "./Brackets";
-import {QueryDeepPartialEntity} from "./QueryPartialEntity";
-import {EntityMetadata} from "../metadata/EntityMetadata";
-import {ColumnMetadata} from "../metadata/ColumnMetadata";
-import {SqljsDriver} from "../driver/sqljs/SqljsDriver";
-import {PostgresDriver} from "../driver/postgres/PostgresDriver";
-import {CockroachDriver} from "../driver/cockroachdb/CockroachDriver";
-import {SqlServerDriver} from "../driver/sqlserver/SqlServerDriver";
-import {OracleDriver} from "../driver/oracle/OracleDriver";
-import {EntitySchema} from "../";
-import {FindOperator} from "../find-options/FindOperator";
-import {In} from "../find-options/operator/In";
-import {EntityColumnNotFound} from "../error/EntityColumnNotFound";
+import {EntityTarget} from "../../common/EntityTarget";
+import {Alias} from "../Alias";
+import {Brackets} from "../Brackets";
+import {QueryDeepPartialEntity} from "../QueryPartialEntity";
+import {EntityMetadata} from "../../metadata/EntityMetadata";
+import {ColumnMetadata} from "../../metadata/ColumnMetadata";
+import {SqljsDriver} from "../../driver/sqljs/SqljsDriver";
+import {PostgresDriver} from "../../driver/postgres/PostgresDriver";
+import {CockroachDriver} from "../../driver/cockroachdb/CockroachDriver";
+import {SqlServerDriver} from "../../driver/sqlserver/SqlServerDriver";
+import {OracleDriver} from "../../driver/oracle/OracleDriver";
+import {EntitySchema} from "../../index";
+import {FindOperator} from "../../find-options/FindOperator";
+import {In} from "../../find-options/operator/In";
+import {EntityColumnNotFound} from "../../error/EntityColumnNotFound";
 
 // todo: completely cover query builder with tests
 // todo: entityOrProperty can be target name. implement proper behaviour if it is.
