+++ src/query-builder/builder/SoftDeleteQueryBuilder.ts
@@ -1,26 +1,26 @@
-import {CockroachDriver} from "../driver/cockroachdb/CockroachDriver";
+import {CockroachDriver} from "../../driver/cockroachdb/CockroachDriver";
 import {QueryBuilder} from "./QueryBuilder";
-import {ObjectLiteral} from "../common/ObjectLiteral";
-import {EntityTarget} from "../common/EntityTarget";
-import {Connection} from "../connection/Connection";
-import {QueryRunner} from "../query-runner/QueryRunner";
-import {SqlServerDriver} from "../driver/sqlserver/SqlServerDriver";
-import {PostgresDriver} from "../driver/postgres/PostgresDriver";
-import {WhereExpression} from "./WhereExpression";
-import {Brackets} from "./Brackets";
-import {UpdateResult} from "./result/UpdateResult";
-import {ReturningStatementNotSupportedError} from "../error/ReturningStatementNotSupportedError";
-import {ReturningResultsEntityUpdator} from "./ReturningResultsEntityUpdator";
-import {SqljsDriver} from "../driver/sqljs/SqljsDriver";
-import {MysqlDriver} from "../driver/mysql/MysqlDriver";
-import {BroadcasterResult} from "../subscriber/BroadcasterResult";
-import {AbstractSqliteDriver} from "../driver/sqlite-abstract/AbstractSqliteDriver";
-import {OrderByCondition} from "../find-options/OrderByCondition";
-import {LimitOnUpdateNotSupportedError} from "../error/LimitOnUpdateNotSupportedError";
-import {MissingDeleteDateColumnError} from "../error/MissingDeleteDateColumnError";
-import {OracleDriver} from "../driver/oracle/OracleDriver";
-import {UpdateValuesMissingError} from "../error/UpdateValuesMissingError";
-import {EntitySchema} from "../entity-schema/EntitySchema";
+import {ObjectLiteral} from "../../common/ObjectLiteral";
+import {EntityTarget} from "../../common/EntityTarget";
+import {Connection} from "../../connection/Connection";
+import {QueryRunner} from "../../query-runner/QueryRunner";
+import {SqlServerDriver} from "../../driver/sqlserver/SqlServerDriver";
+import {PostgresDriver} from "../../driver/postgres/PostgresDriver";
+import {WhereExpression} from "../WhereExpression";
+import {Brackets} from "../Brackets";
+import {UpdateResult} from "../result/UpdateResult";
+import {ReturningStatementNotSupportedError} from "../../error/ReturningStatementNotSupportedError";
+import {ReturningResultsEntityUpdator} from "../ReturningResultsEntityUpdator";
+import {SqljsDriver} from "../../driver/sqljs/SqljsDriver";
+import {MysqlDriver} from "../../driver/mysql/MysqlDriver";
+import {BroadcasterResult} from "../../subscriber/BroadcasterResult";
+import {AbstractSqliteDriver} from "../../driver/sqlite-abstract/AbstractSqliteDriver";
+import {OrderByCondition} from "../../find-options/OrderByCondition";
+import {LimitOnUpdateNotSupportedError} from "../../error/LimitOnUpdateNotSupportedError";
+import {MissingDeleteDateColumnError} from "../../error/MissingDeleteDateColumnError";
+import {OracleDriver} from "../../driver/oracle/OracleDriver";
+import {UpdateValuesMissingError} from "../../error/UpdateValuesMissingError";
+import {EntitySchema} from "../../entity-schema/EntitySchema";
 
 /**
  * Allows to build complex sql queries in a fashion way and execute those queries.
