+++ src/query-builder/builder/UpdateQueryBuilder.ts
@@ -1,30 +1,30 @@
-import {CockroachDriver} from "../driver/cockroachdb/CockroachDriver";
-import {SapDriver} from "../driver/sap/SapDriver";
-import { ColumnMetadata } from "../metadata/ColumnMetadata";
+import {CockroachDriver} from "../../driver/cockroachdb/CockroachDriver";
+import {SapDriver} from "../../driver/sap/SapDriver";
+import { ColumnMetadata } from "../../metadata/ColumnMetadata";
 import {QueryBuilder} from "./QueryBuilder";
-import {ObjectLiteral} from "../common/ObjectLiteral";
-import {Connection} from "../connection/Connection";
-import {QueryRunner} from "../query-runner/QueryRunner";
-import {SqlServerDriver} from "../driver/sqlserver/SqlServerDriver";
-import {PostgresDriver} from "../driver/postgres/PostgresDriver";
-import {WhereExpression} from "./WhereExpression";
-import {Brackets} from "./Brackets";
-import {EntityMetadata} from "../metadata/EntityMetadata";
-import {UpdateResult} from "./result/UpdateResult";
-import {ReturningStatementNotSupportedError} from "../error/ReturningStatementNotSupportedError";
-import {ReturningResultsEntityUpdator} from "./ReturningResultsEntityUpdator";
-import {SqljsDriver} from "../driver/sqljs/SqljsDriver";
-import {MysqlDriver} from "../driver/mysql/MysqlDriver";
-import {BroadcasterResult} from "../subscriber/BroadcasterResult";
-import {AbstractSqliteDriver} from "../driver/sqlite-abstract/AbstractSqliteDriver";
-import {OrderByCondition} from "../find-options/OrderByCondition";
-import {LimitOnUpdateNotSupportedError} from "../error/LimitOnUpdateNotSupportedError";
-import {OracleDriver} from "../driver/oracle/OracleDriver";
-import {UpdateValuesMissingError} from "../error/UpdateValuesMissingError";
-import {EntityColumnNotFound} from "../error/EntityColumnNotFound";
-import {QueryDeepPartialEntity} from "./QueryPartialEntity";
-import {AuroraDataApiDriver} from "../driver/aurora-data-api/AuroraDataApiDriver";
-import {BetterSqlite3Driver} from "../driver/better-sqlite3/BetterSqlite3Driver";
+import {ObjectLiteral} from "../../common/ObjectLiteral";
+import {Connection} from "../../connection/Connection";
+import {QueryRunner} from "../../query-runner/QueryRunner";
+import {SqlServerDriver} from "../../driver/sqlserver/SqlServerDriver";
+import {PostgresDriver} from "../../driver/postgres/PostgresDriver";
+import {WhereExpression} from "../WhereExpression";
+import {Brackets} from "../Brackets";
+import {EntityMetadata} from "../../metadata/EntityMetadata";
+import {UpdateResult} from "../result/UpdateResult";
+import {ReturningStatementNotSupportedError} from "../../error/ReturningStatementNotSupportedError";
+import {ReturningResultsEntityUpdator} from "../ReturningResultsEntityUpdator";
+import {SqljsDriver} from "../../driver/sqljs/SqljsDriver";
+import {MysqlDriver} from "../../driver/mysql/MysqlDriver";
+import {BroadcasterResult} from "../../subscriber/BroadcasterResult";
+import {AbstractSqliteDriver} from "../../driver/sqlite-abstract/AbstractSqliteDriver";
+import {OrderByCondition} from "../../find-options/OrderByCondition";
+import {LimitOnUpdateNotSupportedError} from "../../error/LimitOnUpdateNotSupportedError";
+import {OracleDriver} from "../../driver/oracle/OracleDriver";
+import {UpdateValuesMissingError} from "../../error/UpdateValuesMissingError";
+import {EntityColumnNotFound} from "../../error/EntityColumnNotFound";
+import {QueryDeepPartialEntity} from "../QueryPartialEntity";
+import {AuroraDataApiDriver} from "../../driver/aurora-data-api/AuroraDataApiDriver";
+import {BetterSqlite3Driver} from "../../driver/better-sqlite3/BetterSqlite3Driver";
 
 /**
  * Allows to build complex sql queries in a fashion way and execute those queries.
