+++ src/repository/MongoRepository.ts
@@ -35,7 +35,7 @@ import {
 } from "../driver/mongodb/typings";
 import {MongoEntityManager} from "../entity-manager/MongoEntityManager";
 import {QueryRunner} from "../query-runner/QueryRunner";
-import {SelectQueryBuilder} from "../query-builder/SelectQueryBuilder";
+import {SelectQueryBuilder} from "../query-builder/builder/SelectQueryBuilder";
 
 /**
  * Repository used to manage mongodb documents of a single entity type.
